Use_Case_II/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
h2o==3.44.0.3
idna==3.10
joblib==1.5.2
kagglehub==0.3.13
multidict==6.6.4
numpy==1.26.4
packaging==25.0
pandas==2.2.2
propcache==0.3.2
pyarrow==16.1.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scalar_fastapi==1.3.0
scikit-learn==1.5.2
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.3
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.66.5
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
